Добавлена валидация администратора, и всех его endpoint

+++ src/main/java/com/englishaoe/lesson/controllers/AdminController.java
@@ -2,8 +2,10 @@ package com.englishaoe.lesson.controllers;
 
 import com.englishaoe.lesson.database.entity.variants.Task;
 import com.englishaoe.lesson.database.entity.variants.Variant;
+import com.englishaoe.lesson.database.services.CustomerServices;
 import com.englishaoe.lesson.database.services.VariantService;
 import com.englishaoe.lesson.dto.lesson.TaskDTO;
+import com.englishaoe.lesson.exceptions.RegularException;
 import com.englishaoe.lesson.utility.JwtUtil;
 import com.englishaoe.lesson.utility.file.FileUtil;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -23,15 +25,24 @@ public class AdminController {
     JwtUtil jwtUtil;
     @Autowired
     VariantService variantService;
+    @Autowired
+    CustomerServices customerServices;
     @Value("${image.folderDir}")
     private String imageFolderPath;
-
+    @GetMapping("/validate")
+    public boolean validateAdmin(@RequestHeader("Authorization") String token) {
+        return customerServices.isCustomerHasAdminRole(Long.valueOf(jwtUtil.extractSubject(token)));
+    }
     @PostMapping("/upload-variant")
     public ResponseEntity<Variant> uploadVariant(@RequestHeader("Authorization") String token,
                                                  @RequestParam("variantImg") MultipartFile variantImg,
                                                  @RequestParam("variantName") String variantName,
                                                  @RequestParam("creationDate") String creationDate){
-        jwtUtil.extractSubject(token);//TODO check for admin in future
+
+        //jwtUtil.extractSubject(token);//TODO check for admin in future
+        if (!customerServices.isCustomerHasAdminRole(Long.valueOf(jwtUtil.extractSubject(token))))
+            throw new RegularException("Access denied", HttpStatus.FORBIDDEN.value());
+
         return ResponseEntity.ok(variantService.saveVariant(
                 variantService.assembleVariant(
                     FileUtil.saveFileToDir(variantImg, imageFolderPath),
@@ -41,6 +52,8 @@ public class AdminController {
     @DeleteMapping("/delete-variant/{variantId}")
     public ResponseEntity<String> deleteVariant(@RequestHeader("Authorization") String token,
                                                 @PathVariable("variantId") Long variantId){
+        if (!customerServices.isCustomerHasAdminRole(Long.valueOf(jwtUtil.extractSubject(token))))
+            throw new RegularException("Access denied", HttpStatus.FORBIDDEN.value());
         variantService.deleteVariantById(variantId);
         return ResponseEntity.status(HttpStatus.OK).body("Variant deleted");
     }
@@ -51,7 +64,8 @@ public class AdminController {
                                               @RequestParam("img") MultipartFile secondTaskImage,
                                               @RequestParam("secondImg") MultipartFile fourthTaskImageFirst,
                                               @RequestParam("firstImg") MultipartFile fourthTaskImageSecond) {
-        jwtUtil.extractSubject(token);//TODO check for admin in future
+        if (!customerServices.isCustomerHasAdminRole(Long.valueOf(jwtUtil.extractSubject(token))))
+            throw new RegularException("Access denied", HttpStatus.FORBIDDEN.value());
         tasksJSON = String.format(tasksJSON,
                 FileUtil.saveFileToDir(secondTaskImage, imageFolderPath),
                 FileUtil.saveFileToDir(fourthTaskImageFirst, imageFolderPath),
